Ограничение просмотра некоторых данных пользователям. Климов

+++ Pages/Engine.xaml.cs
@@ -22,6 +22,10 @@ namespace AutoReview.Pages
             InitializeComponent();
             mainWindow = _mainWindow;
             context = new AppDbContext($"server=localhost;port=3307;database=AutoReview;user={AuthData.Login};password={AuthData.Password};");
+            if (AuthData.Rights == false)
+            {
+                id.Visibility = Visibility.Collapsed;
+            }
             LoadData();
         }
 
@@ -32,88 +36,41 @@ namespace AutoReview.Pages
 
         private void AddEngine(object sender, RoutedEventArgs e)
         {
-            var window = new Window
-            {
-                Title = "Добавить двигатель",
-                Height = 300,
-                Width = 400,
-                WindowStartupLocation = WindowStartupLocation.CenterScreen,
-                ResizeMode = ResizeMode.NoResize
-            };
-
-            var editControl = new EngineEditControl
-            {
-                EngineType = "",
-                EngineCapacity = "",
-                EnginePower = "",
-                EngineId = null
-            };
-
-            editControl.OnSave += (control) =>
-            {
-
-                var engine = new Classes.Engine
-                {
-                    Type_Engine = control.EngineType,
-                    Capacity_Engine = decimal.TryParse(control.EngineCapacity, out decimal capacity) ? capacity : 0,
-                    Power_Engine = int.TryParse(control.EnginePower, out int power) ? power : 0
-                };
-
-                context.Engine.Add(engine);
-                context.SaveChanges();
-
-                MessageBox.Show($"Двигатель {engine.Type_Engine} успешно добавлен!");
-
-                window.Close();
-                LoadData();
-
-            };
-
-            editControl.OnCancel += () => window.Close();
-
-            window.Content = editControl;
-            window.ShowDialog();
-        }
-
-        private void EditEngine(object sender, RoutedEventArgs e)
-        {
-            if (engineList.SelectedItem is Classes.Engine selected)
+            if (AuthData.Rights)
             {
                 var window = new Window
                 {
-                    Title = "Редактировать двигатель",
+                    Title = "Добавить двигатель",
+                    Height = 300,
                     Width = 400,
-                    Height = 350,
                     WindowStartupLocation = WindowStartupLocation.CenterScreen,
                     ResizeMode = ResizeMode.NoResize
                 };
 
                 var editControl = new EngineEditControl
                 {
-                    EngineType = selected.Type_Engine,
-                    EngineCapacity = selected.Capacity_Engine.ToString(),
-                    EnginePower = selected.Power_Engine.ToString(),
-                    EngineId = selected.Id_Engine
+                    EngineType = "",
+                    EngineCapacity = "",
+                    EnginePower = "",
+                    EngineId = null
                 };
 
                 editControl.OnSave += (control) =>
                 {
-
-                    var engine = context.Engine.Find(control.EngineId);
-                    
-                    if (engine != null)
+                    var engine = new Classes.Engine
                     {
-                        engine.Type_Engine = control.EngineType;
-                        engine.Capacity_Engine = decimal.TryParse(control.EngineCapacity, out decimal capacity) ? capacity : 0;
-                        engine.Power_Engine = int.TryParse(control.EnginePower, out int power) ? power : 0;
+                        Type_Engine = control.EngineType,
+                        Capacity_Engine = decimal.TryParse(control.EngineCapacity, out decimal capacity) ? capacity : 0,
+                        Power_Engine = int.TryParse(control.EnginePower, out int power) ? power : 0
+                    };
 
-                        context.SaveChanges();
-                        MessageBox.Show("Двигатель обновлен!");
+                    context.Engine.Add(engine);
+                    context.SaveChanges();
 
-                        window.Close();
-                        LoadData();
-                    }
+                    MessageBox.Show($"Двигатель {engine.Type_Engine} успешно добавлен!");
 
+                    window.Close();
+                    LoadData();
                 };
 
                 editControl.OnCancel += () => window.Close();
@@ -122,31 +79,98 @@ namespace AutoReview.Pages
             }
             else
             {
-                MessageBox.Show("Выберите двигатель для редактирования!");
+                MessageBox.Show("Вы не можете добавлять данные!");
+            }
+            
+        }
+
+        private void EditEngine(object sender, RoutedEventArgs e)
+        {
+            if (AuthData.Rights)
+            {
+                
+
+                if (engineList.SelectedItem is Classes.Engine selected)
+                {
+                    var window = new Window
+                    {
+                        Title = "Редактировать двигатель",
+                        Width = 400,
+                        Height = 350,
+                        WindowStartupLocation = WindowStartupLocation.CenterScreen,
+                        ResizeMode = ResizeMode.NoResize
+                    };
+
+                    var editControl = new EngineEditControl
+                    {
+                        EngineType = selected.Type_Engine,
+                        EngineCapacity = selected.Capacity_Engine.ToString(),
+                        EnginePower = selected.Power_Engine.ToString(),
+                        EngineId = selected.Id_Engine
+                    };
+
+                    editControl.OnSave += (control) =>
+                    {
+
+                        var engine = context.Engine.Find(control.EngineId);
+
+                        if (engine != null)
+                        {
+                            engine.Type_Engine = control.EngineType;
+                            engine.Capacity_Engine = decimal.TryParse(control.EngineCapacity, out decimal capacity) ? capacity : 0;
+                            engine.Power_Engine = int.TryParse(control.EnginePower, out int power) ? power : 0;
+
+                            context.SaveChanges();
+                            MessageBox.Show("Двигатель обновлен!");
+
+                            window.Close();
+                            LoadData();
+                        }
+
+                    };
+
+                    editControl.OnCancel += () => window.Close();
+                    window.Content = editControl;
+                    window.ShowDialog();
+                }
+                else
+                {
+                    MessageBox.Show("Выберите двигатель для редактирования!");
+                }
+            }
+            else
+            {
+                MessageBox.Show("Вы не можете редактировать данные!");
             }
         }
 
         private void DeleteEngine(object sender, RoutedEventArgs e)
         {
-            if (engineList.SelectedItem is Classes.Engine selected)
+            if (AuthData.Rights)
             {
-                if (MessageBox.Show($"Удалить двигатель {selected.Type_Engine}?",
-                    "Подтверждение", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+                if (engineList.SelectedItem is Classes.Engine selected)
                 {
-                    var engine = context.Engine.Find(selected.Id_Engine);
-
-                    if (engine != null)
+                    if (MessageBox.Show($"Удалить двигатель {selected.Type_Engine}?", "Подтверждение", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
                     {
-                        context.Engine.Remove(engine);
-                        context.SaveChanges();
-                        MessageBox.Show("Двигатель удален!");
-                        LoadData();
+                        var engine = context.Engine.Find(selected.Id_Engine);
+
+                        if (engine != null)
+                        {
+                            context.Engine.Remove(engine);
+                            context.SaveChanges();
+                            MessageBox.Show("Двигатель удален!");
+                            LoadData();
+                        }
                     }
+                }
+                else
+                {
+                    MessageBox.Show("Выберите двигатель для удаления!", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
                 }
             }
             else
             {
-                MessageBox.Show("Выберите двигатель для удаления!");
+                MessageBox.Show("Вы не можете удалять данные!");
             }
         }
